# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ComponentStatus": {
        "enum": [
          "ok",
          "degraded",
          "down",
          "skipped"
        ],
        "title": "ComponentStatus",
        "type": "string"
      },
      "DependencyHealth": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ComponentStatus"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "DependencyHealth",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LivenessResponse": {
        "properties": {
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ComponentStatus",
            "default": "ok"
          },
          "uptime_s": {
            "description": "Seconds since process start",
            "title": "Uptime S",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "service",
          "version",
          "uptime_s"
        ],
        "title": "LivenessResponse",
        "type": "object"
      },
      "ReadinessResponse": {
        "properties": {
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyHealth"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "profile": {
            "title": "Profile",
            "type": "string"
          },
          "service": {
            "title": "Service",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ComponentStatus"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "status",
          "service",
          "version",
          "profile",
          "dependencies"
        ],
        "title": "ReadinessResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Agentic Digital Twin Platform for Aircraft Engine Predictive Maintenance.\n\nReplays the NASA C-MAPSS turbofan degradation dataset as live telemetry, maintains\nan event-sourced digital twin per engine, predicts Remaining Useful Life with deep\nmodels, detects anomalies, explains predictions, and reasons over the fleet with a\nLangGraph multi-agent system.",
    "title": "AeroTwin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/engines/{engine_ref}": {
      "get": {
        "description": "Full twin state. Accepts a UUID, a unit number, or an external ref.",
        "operationId": "get_engine_api_v1_engines__engine_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "engine_ref",
            "required": true,
            "schema": {
              "title": "Engine Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Engine Api V1 Engines  Engine Ref  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Twin detail",
        "tags": [
          "fleet",
          "engines"
        ]
      }
    },
    "/api/v1/engines/{engine_ref}/commands/{command}": {
      "post": {
        "description": "Queue a control command. Asynchronous by design (single-writer rule).",
        "operationId": "send_command_api_v1_engines__engine_ref__commands__command__post",
        "parameters": [
          {
            "in": "path",
            "name": "engine_ref",
            "required": true,
            "schema": {
              "title": "Engine Ref",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "command",
            "required": true,
            "schema": {
              "title": "Command",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Send Command Api V1 Engines  Engine Ref  Commands  Command  Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Send Command",
        "tags": [
          "fleet",
          "replay"
        ]
      }
    },
    "/api/v1/engines/{engine_ref}/explain": {
      "get": {
        "description": "Per-sensor attribution for the current prediction (Doc 07 section 7.7).\n\nComputed on demand rather than every cycle: the gradient pass roughly\ndoubles inference cost, and an explanation is only read when a user opens\nthe panel.",
        "operationId": "get_explanation_api_v1_engines__engine_ref__explain_get",
        "parameters": [
          {
            "in": "path",
            "name": "engine_ref",
            "required": true,
            "schema": {
              "title": "Engine Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Explanation Api V1 Engines  Engine Ref  Explain Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Attribution",
        "tags": [
          "fleet",
          "predictions"
        ]
      }
    },
    "/api/v1/engines/{engine_ref}/history": {
      "get": {
        "description": "Time series for the engine detail charts.\n\nDecimated server-side: a chart plotting 200 points has no use for 600\nsamples, and shipping them would triple the payload for pixels that overlap.",
        "operationId": "get_history_api_v1_engines__engine_ref__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "engine_ref",
            "required": true,
            "schema": {
              "title": "Engine Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 600,
              "minimum": 10,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "from_cycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Cycle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get History Api V1 Engines  Engine Ref  History Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Chart history",
        "tags": [
          "fleet",
          "engines"
        ]
      }
    },
    "/api/v1/fleet": {
      "get": {
        "description": "Paginated fleet listing with sorting and filtering.",
        "operationId": "list_fleet_api_v1_fleet_get",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "health",
              "enum": [
                "health",
                "rul",
                "cycle",
                "unit",
                "priority"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "asc",
              "enum": [
                "asc",
                "desc"
              ],
              "title": "Order",
              "type": "string"
            }
          },
          {
            "description": "Comma-separated health bands",
            "in": "query",
            "name": "band",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated health bands",
              "title": "Band"
            }
          },
          {
            "description": "Match tail number or unit",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Match tail number or unit",
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Fleet Api V1 Fleet Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "List the fleet",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/fleet/summary": {
      "get": {
        "operationId": "fleet_summary_endpoint_api_v1_fleet_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Fleet Summary Endpoint Api V1 Fleet Summary Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Fleet KPIs",
        "tags": [
          "fleet"
        ]
      }
    },
    "/api/v1/system": {
      "get": {
        "operationId": "system_stats_api_v1_system_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response System Stats Api V1 System Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Engine runtime statistics",
        "tags": [
          "fleet",
          "admin"
        ]
      }
    },
    "/health/deep": {
      "get": {
        "description": "Expose runtime configuration and subsystem state for operators.\n\nSecrets are never included -- only whether they are configured.",
        "operationId": "deep_health_deep_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Deep Health Deep Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Deep diagnostic",
        "tags": [
          "health"
        ]
      }
    },
    "/health/live": {
      "get": {
        "description": "Return 200 whenever the process is running and the event loop is responsive.",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LivenessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Liveness probe",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "description": "Check every backing dependency.\n\nIn M1 the datastores are not yet wired, so they report ``skipped`` rather\nthan faking success. M2 replaces these with real connection checks.",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "summary": "Readiness probe",
        "tags": [
          "health"
        ]
      }
    }
  },
  "tags": [
    {
      "description": "Liveness, readiness and deep diagnostics.",
      "name": "health"
    },
    {
      "description": "Fleet-wide listing, ranking and aggregates.",
      "name": "fleet"
    },
    {
      "description": "Individual twin state, telemetry and timeline.",
      "name": "engines"
    },
    {
      "description": "RUL predictions and explanations.",
      "name": "predictions"
    },
    {
      "description": "Anomaly detections and acknowledgement.",
      "name": "anomalies"
    },
    {
      "description": "Replay clock control (asynchronous commands).",
      "name": "replay"
    },
    {
      "description": "What-if scenario simulation.",
      "name": "simulate"
    },
    {
      "description": "Conversational engineering assistant.",
      "name": "copilot"
    },
    {
      "description": "Agent runs, traces and topology.",
      "name": "agents"
    },
    {
      "description": "RAG corpus search and documents.",
      "name": "knowledge"
    },
    {
      "description": "Work packages and scheduling.",
      "name": "maintenance"
    },
    {
      "description": "Operator tooling and system state.",
      "name": "admin"
    }
  ]
}